moved ObjectCreationHandler from /handler. Interfaces to /handlers
